refactor: remove legacy future_state family and dead shared_state methods
Drops the parallel "old" state-machine that was kept alive only by tests
and the generator's base class:

* Delete `future_state<T>` template, `future_state_base`, `future_state_enum`,
  `future_states::` constants, `intrusive_ptr_add_ref/release` overloads
  for the base, plus `detail/future_state.hpp` and `impl/detail/future_state.ipp`.
* Extract `result_storage<T>` / `result_storage<void>` and the trivially-
  move-and-destructible trait into a new `detail/result_storage.hpp`;
  `shared_state.hpp` switches its include.
* `generator_state` is now standalone (no inheritance from `future_state_base`):
  it owns its own refcount, atomic state byte, owning/resume coroutine
  handles, and `owns_coroutine_` flag; manual `add_ref()`/`release()` calls
  remain unchanged (the type is held everywhere via raw `generator_state<T>*`,
  there were no `intrusive_ptr<generator_state>` or polymorphic uses).
* Remove zero-caller dead methods `shared_state::is_future_released()` and
  `shared_state::deallocate_from_finalizer()` (both T and void specs).
* Remove the empty backward-compat shim `mailbox/message_result.hpp` and
  its lone include from `mailbox.hpp` (`default_mailbox.hpp` already pulls
  in `message.hpp` transitively).
* `forwards.hpp` drops the now-deleted forward decls; `src.hpp` drops the
  now-empty `future_state.ipp` include.

Legacy tests deleted (they tested the removed `future_state<T>` directly):
`test/slot-refcount/`, `test/continuation/`, and the `future_state<int/void>`
sections in `test/coroutines/main.cpp:144-241`. The release-acquire
side-effect property tested by `slot-refcount` is preserved by porting it
to `shared_state` in `test/future-state-fixes/main.cpp` ("Concurrent:
is_ready acquire synchronizes side effect on shared_state"); the basic
cross-thread value-visibility property is already covered there.
