--- header/actor-zeta/mailbox/message_result.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-// This header is now empty - implementations moved to message.hpp and message.ipp
-// Kept for backwards compatibility (in case existing code includes it)
-
-#include <actor-zeta/detail/future.hpp>
-#include <actor-zeta/detail/generator.hpp>
-#include <actor-zeta/mailbox/message.hpp>
