+++ header/actor-zeta/src.hpp
@@ -7,7 +7,6 @@
 
 #include <actor-zeta/impl/actor/address.ipp>
 
-#include <actor-zeta/impl/detail/future_state.ipp>
 #include <actor-zeta/impl/detail/launch_thread.ipp>
 #include <actor-zeta/impl/detail/ref_counted.ipp>
