+++ test/CMakeLists.txt
@@ -4,7 +4,6 @@ add_subdirectory(tooltestsuites)
 
 add_subdirectory(actor-id)
 add_subdirectory(bool-prohibited)
-add_subdirectory(continuation)
 add_subdirectory(coroutines)
 add_subdirectory(coroutine-threading)
 add_subdirectory(dispatch-traits)
@@ -22,7 +21,6 @@ add_subdirectory(rtt)
 add_subdirectory(resume-info)
 add_subdirectory(type-validation)
 add_subdirectory(shutdown)
-add_subdirectory(slot-refcount)
 add_subdirectory(coroutine-allocation)
 add_subdirectory(future-state-fixes)
 add_subdirectory(rtt-issues)
