+++ test/coroutines/main.cpp
@@ -5,7 +5,6 @@
 #include <actor-zeta.hpp>
 #include <actor-zeta/actor/dispatch.hpp>
 #include <actor-zeta/config.hpp>
-#include <actor-zeta/detail/future_state.hpp>
 
 // ============================================================================
 // Test 1: promise_type exists and can be used for co_return
@@ -133,121 +132,6 @@ TEST_CASE("simple coroutines with co_return") {
         std::move(future).take_ready();  // Should not throw
     }
 }
-
-// ============================================================================
-// Test 3: future_state has coroutine storage and methods
-// ============================================================================
-
-TEST_CASE("future_state coroutine methods") {
-    auto* resource =std::pmr::get_default_resource();
-
-    SECTION("future_state<int> has coroutine methods") {
-        void* mem = resource->allocate(sizeof(actor_zeta::detail::future_state<int>),
-                                        alignof(actor_zeta::detail::future_state<int>));
-        auto* state = new (mem) actor_zeta::detail::future_state<int>(resource);
-
-        // Test resume_coroutine exists and can be called
-        state->resume_coroutine();  // Should not crash (no-op if no coroutine)
-
-        // Cleanup
-        state->release();
-    }
-
-    SECTION("future_state<void> has coroutine methods") {
-        void* mem = resource->allocate(sizeof(actor_zeta::detail::future_state<void>),
-                                        alignof(actor_zeta::detail::future_state<void>));
-        auto* state = new (mem) actor_zeta::detail::future_state<void>(resource);
-
-        // Test resume_coroutine exists
-        state->resume_coroutine();
-
-        // Cleanup
-        state->release();
-    }
-}
-
-// ============================================================================
-// Test 4: Coroutine handle storage via set_coroutine
-// ============================================================================
-
-TEST_CASE("coroutine handle can be stored in future_state") {
-    auto* resource =std::pmr::get_default_resource();
-    auto actor = actor_zeta::spawn<coroutine_test_actor>(resource);
-
-    SECTION("set_coroutine stores handle") {
-        void* mem = resource->allocate(sizeof(actor_zeta::detail::future_state<int>),
-                                        alignof(actor_zeta::detail::future_state<int>));
-        auto* state = new (mem) actor_zeta::detail::future_state<int>(resource);
-
-        // FIXED: Use send() instead of direct call
-        auto [needs_sched, future] = actor_zeta::send(
-            actor.get(),
-            &coroutine_test_actor::coro_storage
-        );
-
-        // Process message
-        actor->resume(100);
-
-        // Test resume_coroutine exists (no-op if no coroutine stored)
-        state->resume_coroutine();
-
-        // Cleanup
-        state->release();
-    }
-}
-
-// ============================================================================
-// Test 5: Virtual methods with `final` keyword (devirtualization check)
-// ============================================================================
-
-TEST_CASE("virtual methods marked as final") {
-    // This is a compile-time check - if it compiles, final is working
-    auto* resource =std::pmr::get_default_resource();
-
-    void* mem = resource->allocate(sizeof(actor_zeta::detail::future_state<int>),
-                                    alignof(actor_zeta::detail::future_state<int>));
-    auto* state = new (mem) actor_zeta::detail::future_state<int>(resource);
-
-    // Cast to base to ensure virtual call
-    actor_zeta::detail::future_state_base* base = state;
-
-    // Test resume_coroutine exists (no-op if no coroutine stored)
-    base->resume_coroutine();
-
-    // Cleanup
-    state->release();
-}
-
-// ============================================================================
-// Test 6: promise_type creates future_state correctly
-// ============================================================================
-
-// ============================================================================
-// Test 7: Coroutine destruction
-// ============================================================================
-
-TEST_CASE("future_state destroys stored coroutine") {
-    auto* resource =std::pmr::get_default_resource();
-
-    SECTION("destructor calls coro_handle_.destroy() if not done") {
-        // Create future_state
-        void* mem = resource->allocate(sizeof(actor_zeta::detail::future_state<int>),
-                                        alignof(actor_zeta::detail::future_state<int>));
-        auto* state = new (mem) actor_zeta::detail::future_state<int>(resource);
-
-        // Destroy immediately (no coroutine stored - should be safe)
-        state->~future_state();
-        resource->deallocate(state, sizeof(actor_zeta::detail::future_state<int>),
-                            alignof(actor_zeta::detail::future_state<int>));
-
-        // If we reach here, no crash occurred
-        REQUIRE(true);
-    }
-}
-
-// ============================================================================
-// Tests 8-11: REMOVED - behavior_t is outdated
-// ============================================================================
 
 // ============================================================================
 // Test 12: Coroutine futures (STATE mode)
